Подправил вывод в консоль. Не реализована  правильная очередность  строительства.

+++ HouseBuilding/IPart.cs
@@ -6,11 +6,11 @@ namespace HouseBuilding
 {
     public interface IPart
     {
-     
-        decimal PercentCompleted { get; set; }
-       
+
+        int PercentCompleted { get; set; }
+
         bool Completed { get; set; }
         int Building();
-        
+
     }
 }
